Backend-Documents/Code Mẫu/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary  # Driver cho Neon/Postgres
django-cors-headers